requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
asgiref==3.5.2
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
Django==4.1.4
fastapi==0.87.0
filetype==1.2.0
flatbuffers==22.12.6
frozenlist==1.3.3
h11==0.14.0
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.22.4
llvmlite==0.39.1
mpmath==1.2.1
multidict==6.0.3
networkx==2.8.8
numba==0.56.4
numpy==1.23.5
onnxruntime==1.13.1
opencv-python-headless==4.6.0.66
packaging==22.0
Pillow==9.3.0
pooch==1.6.0
protobuf==4.21.12
pydantic==1.10.2
PyMatting==1.1.8
python-multipart==0.0.5
PyWavelets==1.4.1
rembg==2.0.29
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.3
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
starlette==0.21.0
sympy==1.11.1
tifffile==2022.10.10
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
watchdog==2.1.9
yarl==1.8.2